+++ XInputBridge/GlobalState.cpp
@@ -9,14 +9,13 @@
 #include <spdlog/sinks/basic_file_sink.h>
 
 #include "Types.h"
-#include "UniUtil.h"
-#include "DsHidMini/dshmguid.h"
+
 
 EXTERN_C IMAGE_DOS_HEADER __ImageBase;
 
 void GlobalState::Initialize()
 {
-	InitializeCriticalSection(&StatesLock);
+	InitializeSRWLock(&StatesLock);
 
 	CHAR dllPath[MAX_PATH];
 
@@ -65,97 +64,6 @@ void GlobalState::Destroy()
 	(void)CM_Unregister_Notification(XusbNotificationHandle);
 
 	(void)hid_exit();
-
-	DeleteCriticalSection(&StatesLock);
-}
-
-DWORD GlobalState::DeviceNotificationCallback(
-	HCMNOTIFICATION hNotify,
-	PVOID Context,
-	CM_NOTIFY_ACTION Action,
-	PCM_NOTIFY_EVENT_DATA EventData,
-	DWORD EventDataSize
-)
-{
-	UNREFERENCED_PARAMETER(hNotify);
-	UNREFERENCED_PARAMETER(EventDataSize);
-
-	const GlobalState* _this = static_cast<GlobalState*>(Context);
-	const std::shared_ptr<spdlog::logger> logger = spdlog::get(LOGGER_NAME)->clone(__FUNCTION__);
-
-	switch (Action)
-	{
-	case CM_NOTIFY_ACTION_DEVICEINTERFACEARRIVAL:
-		if (IsEqualGUID(GUID_DEVINTERFACE_DSHIDMINI, EventData->u.DeviceInterface.ClassGuid))
-		{
-			logger->info("New DS3 device arrived: {}", ConvertWideToANSI(EventData->u.DeviceInterface.SymbolicLink));
-		}
-
-		if (IsEqualGUID(XUSB_INTERFACE_CLASS_GUID, EventData->u.DeviceInterface.ClassGuid))
-		{
-			const auto symlink = ConvertWideToANSI(EventData->u.DeviceInterface.SymbolicLink);
-
-			logger->info("New XUSB device arrived: {}", symlink);
-
-			DWORD userIndex = INVALID_X_INPUT_USER_ID;
-			if (SymlinkToUserIndex(EventData->u.DeviceInterface.SymbolicLink, &userIndex))
-			{
-				logger->info("User index: {}", userIndex);
-
-				/*if (const auto slot = GetFreeSlot())
-				{
-					slot->SetOnlineAsXusb(symlink, userIndex);
-				}
-				else
-				{
-					logger->warn("No free slot for {}", symlink);
-				}*/
-			}
-			else
-			{
-				logger->error("User index lookup failed");
-			}
-		}
-		break;
-	case CM_NOTIFY_ACTION_DEVICEINTERFACEREMOVAL:
-		if (IsEqualGUID(GUID_DEVINTERFACE_DSHIDMINI, EventData->u.DeviceInterface.ClassGuid))
-		{
-			const std::string symlink = ConvertWideToANSI(EventData->u.DeviceInterface.SymbolicLink);
-			logger->info("DS3 device got removed: {}", symlink);
-
-			/*const auto item = std::ranges::find_if(G_DEVICE_STATES, [symlink](const XI_DEVICE_STATE& element)
-			{
-				return absl::EqualsIgnoreCase(element.SymbolicLink, symlink);
-			});
-
-			if (item != G_DEVICE_STATES.end())
-			{
-				item->Type = XI_DEVICE_TYPE_NOT_CONNECTED;
-			}*/
-		}
-
-		if (IsEqualGUID(XUSB_INTERFACE_CLASS_GUID, EventData->u.DeviceInterface.ClassGuid))
-		{
-			const std::string symlink = ConvertWideToANSI(EventData->u.DeviceInterface.SymbolicLink);
-			logger->info("XUSB device got removed: {}", symlink);
-
-			/*const auto item = std::ranges::find_if(G_DEVICE_STATES, [symlink](const XI_DEVICE_STATE& element)
-			{
-				return absl::EqualsIgnoreCase(element.SymbolicLink, symlink);
-			});
-
-			if (item != G_DEVICE_STATES.end())
-			{
-				item->Type = XI_DEVICE_TYPE_NOT_CONNECTED;
-				item->Backend.UserIndex = K_INVALID_X_INPUT_USER_ID;
-			}*/
-		}
-		break;
-	default:
-		return ERROR_SUCCESS;
-	}
-
-	return ERROR_SUCCESS;
 }
 
 //
@@ -235,113 +143,13 @@ bool GlobalState::SymlinkToUserIndex(PCWSTR Symlink, PDWORD UserIndex)
 	return true;
 }
 
-DWORD WINAPI GlobalState::InitAsync(LPVOID lpParameter)
+DeviceState* GlobalState::GetFreeState()
 {
-	const auto _this = static_cast<GlobalState*>(lpParameter);
-	const std::shared_ptr<spdlog::logger> logger = spdlog::get(LOGGER_NAME)->clone(__FUNCTION__);
+	const auto item = std::ranges::find_if(this->States,
+		[](const DeviceState& element)
+		{
+			return element.Type == XI_DEVICE_TYPE_NOT_CONNECTED;
+		});
 
-	logger->info("Async library startup initialized");
-
-	CHAR systemDir[MAX_PATH] = {};
-
-	if (GetSystemDirectoryA(systemDir, MAX_PATH) == 0)
-	{
-		logger->error("GetSystemDirectoryA failed: {:#x}", GetLastError());
-		return GetLastError();
-	}
-
-	CHAR fullXiPath[MAX_PATH] = {};
-
-	if (PathCombineA(fullXiPath, systemDir, XI_SYSTEM_LIB_NAME) == nullptr)
-	{
-		logger->error("PathCombineA failed: {:#x}", GetLastError());
-		return GetLastError();
-	}
-
-	const HMODULE xiLib = LoadLibraryA(fullXiPath);
-
-	if (xiLib == nullptr)
-	{
-		logger->error("LoadLibraryA failed: {:#x}", GetLastError());
-		return GetLastError();
-	}
-
-	//
-	// Grab the function pointers from the OS-provided exports
-	// 
-
-	_this->FpnXInputGetState = reinterpret_cast<decltype(XInputGetState)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputGetState)
-	));
-	_this->FpnXInputSetState = reinterpret_cast<decltype(XInputSetState)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputSetState)
-	));
-	_this->FpnXInputGetCapabilities = reinterpret_cast<decltype(XInputGetCapabilities)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputGetCapabilities)
-	));
-	_this->FpnXInputEnable = reinterpret_cast<decltype(XInputEnable)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputEnable)
-	));
-	_this->FpnXInputGetDSoundAudioDeviceGuids = reinterpret_cast<decltype(XInputGetDSoundAudioDeviceGuids)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputGetDSoundAudioDeviceGuids)
-	));
-	_this->FpnXInputGetBatteryInformation = reinterpret_cast<decltype(XInputGetBatteryInformation)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputGetBatteryInformation)
-	));
-	_this->FpnXInputGetKeystroke = reinterpret_cast<decltype(XInputGetKeystroke)*>(GetProcAddress(xiLib,
-		NAMEOF(XInputGetKeystroke)
-	));
-	_this->FpnXInputGetStateEx = reinterpret_cast<decltype(XInputGetStateEx)*>(GetProcAddress(xiLib,
-		MAKEINTRESOURCEA(100)
-	));
-	_this->FpnXInputWaitForGuideButton = reinterpret_cast<decltype(XInputWaitForGuideButton)*>(GetProcAddress(xiLib,
-		MAKEINTRESOURCEA(101)
-	));
-	_this->FpnXInputCancelGuideButtonWait = reinterpret_cast<decltype(XInputCancelGuideButtonWait)*>(GetProcAddress(xiLib,
-		MAKEINTRESOURCEA(102)
-	));
-	_this->FpnXInputPowerOffController = reinterpret_cast<decltype(XInputPowerOffController)*>(GetProcAddress(xiLib,
-		MAKEINTRESOURCEA(103)
-	));
-
-	//
-	// Register notifications for device arrival/removal
-	// 
-
-	CM_NOTIFY_FILTER ds3Filter = {};
-	ds3Filter.cbSize = sizeof(CM_NOTIFY_FILTER);
-	ds3Filter.FilterType = CM_NOTIFY_FILTER_TYPE_DEVICEINTERFACE;
-	ds3Filter.u.DeviceInterface.ClassGuid = GUID_DEVINTERFACE_DSHIDMINI;
-
-	//
-	// Register DsHidMini device interface
-	// 
-	CONFIGRET ret = CM_Register_Notification(&ds3Filter, nullptr, DeviceNotificationCallback, &_this->Ds3NotificationHandle);
-
-	if (ret != CR_SUCCESS)
-	{
-		logger->error("CM_Register_Notification (DS3) failed: {:#x}", ret);
-	}
-
-	CM_NOTIFY_FILTER xusbFilter = {};
-	xusbFilter.cbSize = sizeof(CM_NOTIFY_FILTER);
-	xusbFilter.FilterType = CM_NOTIFY_FILTER_TYPE_DEVICEINTERFACE;
-	xusbFilter.u.DeviceInterface.ClassGuid = XUSB_INTERFACE_CLASS_GUID;
-
-	//
-	// Register X360/XBONE device interface
-	// 
-	ret = CM_Register_Notification(&xusbFilter, nullptr, DeviceNotificationCallback, &_this->XusbNotificationHandle);
-
-	if (ret != CR_SUCCESS)
-	{
-		logger->error("CM_Register_Notification (XUSB) failed: {:#x}", ret);
-	}
-
-	if (const auto result = hid_init(); result != 0)
-	{
-		logger->error("hid_init failed: {}", ConvertWideToANSI(hid_error(nullptr)));
-	}
-
-	return ERROR_SUCCESS;
+	return (item != this->States.end()) ? &(*item) : nullptr;
 }
