+++ XInputBridge/XInputBridge.vcxproj
@@ -500,6 +500,7 @@
   <ItemGroup>
     <ClCompile Include="DeviceState.cpp" />
     <ClCompile Include="dllmain.cpp" />
+    <ClCompile Include="GlobalState.Callbacks.cpp" />
     <ClCompile Include="GlobalState.cpp" />
     <ClCompile Include="GlobalState.XInput.cpp" />
     <ClCompile Include="XInputBridge.cpp" />
