+++ XInputBridge/XInputBridge.vcxproj.filters
@@ -64,6 +64,9 @@
     <ClCompile Include="GlobalState.XInput.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="GlobalState.Callbacks.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="XInputBridge.rc">
